Move action network utils
