+++ rebel_management_utilities/members.py
@@ -3,7 +3,7 @@ import datetime
 import pandas as pd
 
 from local_group_support.config.config import get_config
-from local_group_support.utils.action_network import get_forms, query, query_all
+from rebel_management_utilities.action_network import get_forms, query, query_all
 
 FORMATION_DATE = datetime.date(2018, 4, 1)
